move validations interfaces to InventoryApi
