+++ app/code/Magento/Inventory/Model/Source/Validator/CountryValidator.php
@@ -10,6 +10,7 @@ namespace Magento\Inventory\Model\Source\Validator;
 use Magento\Framework\Validation\ValidationResult;
 use Magento\Framework\Validation\ValidationResultFactory;
 use Magento\InventoryApi\Api\Data\SourceInterface;
+use Magento\InventoryApi\Model\SourceValidatorInterface;
 
 /**
  * Check that country id is valid
