+++ app/code/Magento/Inventory/Model/SourceItem/Validator/SourceCodeValidator.php
@@ -10,6 +10,7 @@ namespace Magento\Inventory\Model\SourceItem\Validator;
 use Magento\Framework\Validation\ValidationResult;
 use Magento\Framework\Validation\ValidationResultFactory;
 use Magento\InventoryApi\Api\Data\SourceItemInterface;
+use Magento\InventoryApi\Model\SourceItemValidatorInterface;
 
 /**
  * Check that source code is valid
