+++ app/code/Magento/Inventory/Model/SourceItem/Validator/SourceItemsValidator.php
@@ -10,6 +10,7 @@ namespace Magento\Inventory\Model\SourceItem\Validator;
 use Magento\Framework\Validation\ValidationResult;
 use Magento\Framework\Validation\ValidationResultFactory;
 use Magento\InventoryApi\Api\Data\SourceItemInterface;
+use Magento\InventoryApi\Model\SourceItemValidatorInterface;
 
 /**
  * Responsible for Source items validation
