+++ app/code/Magento/Inventory/etc/di.xml
@@ -17,8 +17,7 @@
     <preference for="Magento\Inventory\Model\Source\Command\GetInterface" type="Magento\Inventory\Model\Source\Command\Get"/>
     <preference for="Magento\Inventory\Model\Source\Command\GetListInterface" type="Magento\Inventory\Model\Source\Command\GetList"/>
     <preference for="Magento\Inventory\Model\Source\Command\SaveInterface" type="Magento\Inventory\Model\Source\Command\Save"/>
-    <preference for="Magento\Inventory\Model\Source\Validator\SourceValidatorInterface" type="Magento\Inventory\Model\Source\Validator\ValidatorChain"/>
-    <type name="Magento\Inventory\Model\Source\Validator\ValidatorChain">
+    <type name="Magento\InventoryApi\Model\SourceValidatorChain">
         <arguments>
             <argument name="validators" xsi:type="array">
                 <item name="name" xsi:type="object">Magento\Inventory\Model\Source\Validator\NameValidator</item>
@@ -36,8 +35,7 @@
     <preference for="Magento\InventoryApi\Api\Data\SourceItemInterface" type="Magento\Inventory\Model\SourceItem"/>
     <preference for="Magento\InventoryApi\Api\Data\SourceItemSearchResultsInterface" type="Magento\Inventory\Model\SourceItemSearchResults"/>
     <preference for="Magento\Inventory\Model\SourceItem\Command\GetListInterface" type="Magento\Inventory\Model\SourceItem\Command\GetList"/>
-    <preference for="Magento\Inventory\Model\SourceItem\Validator\SourceItemValidatorInterface" type="Magento\Inventory\Model\SourceItem\Validator\ValidatorChain"/>
-    <type name="Magento\Inventory\Model\SourceItem\Validator\ValidatorChain">
+    <type name="Magento\InventoryApi\Model\SourceItemValidatorChain">
         <arguments>
             <argument name="validators" xsi:type="array">
                 <item name="sku" xsi:type="object">Magento\Inventory\Model\SourceItem\Validator\SkuValidator</item>
@@ -56,8 +54,7 @@
     <preference for="Magento\Inventory\Model\Stock\Command\GetListInterface" type="Magento\Inventory\Model\Stock\Command\GetList"/>
     <preference for="Magento\Inventory\Model\Stock\Command\GetProductQuantityInterface" type="Magento\Inventory\Model\Stock\Command\GetProductQuantity"/>
     <preference for="Magento\Inventory\Model\Stock\Command\SaveInterface" type="Magento\Inventory\Model\Stock\Command\Save"/>
-    <preference for="Magento\Inventory\Model\Stock\Validator\StockValidatorInterface" type="Magento\Inventory\Model\Stock\Validator\ValidatorChain"/>
-    <type name="Magento\Inventory\Model\Stock\Validator\ValidatorChain">
+    <type name="Magento\InventoryApi\Model\StockValidatorChain">
         <arguments>
             <argument name="validators" xsi:type="array">
                 <item name="name" xsi:type="object">Magento\Inventory\Model\Stock\Validator\NameValidator</item>
@@ -71,8 +68,7 @@
     <preference for="Magento\InventoryApi\Api\StockSourceLinksSaveInterface" type="Magento\Inventory\Model\StockSourceLink\Command\StockSourceLinksSave"/>
     <preference for="Magento\InventoryApi\Api\StockSourceLinksDeleteInterface" type="Magento\Inventory\Model\StockSourceLink\Command\StockSourceLinksDelete"/>
     <preference for="Magento\InventoryApi\Api\UnassignSourceFromStockInterface" type="Magento\Inventory\Model\StockSourceLink\Command\UnassignSourceFromStock"/>
-    <preference for="Magento\InventoryApi\Model\StockSourceLinkValidatorInterface" type="Magento\Inventory\Model\StockSourceLink\Validator\ValidatorChain"/>
-    <type name="Magento\Inventory\Model\StockSourceLink\Validator\ValidatorChain">
+    <type name="Magento\InventoryApi\Model\StockSourceLinkValidatorChain">
         <arguments>
             <argument name="validators" xsi:type="array">
                 <item name="source_code" xsi:type="object">Magento\Inventory\Model\StockSourceLink\Validator\SourceCodeValidator</item>
@@ -83,7 +79,7 @@
     </type>
     <type name="Magento\Inventory\Model\StockSourceLink\Validator\StockSourceLinksValidator">
         <arguments>
-            <argument name="stockSourceLinkValidator" xsi:type="object">Magento\Inventory\Model\StockSourceLink\Validator\ValidatorChain</argument>
+            <argument name="stockSourceLinkValidator" xsi:type="object">Magento\InventoryApi\Model\StockSourceLinkValidatorChain</argument>
         </arguments>
     </type>
 </config>
