+++ app/code/Magento/InventoryApi/Model/StockSourceLinkValidatorChain.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\Inventory\Model\StockSourceLink\Validator;
+namespace Magento\InventoryApi\Model;
 
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Validation\ValidationResult;
@@ -15,8 +15,10 @@ use Magento\InventoryApi\Model\StockSourceLinkValidatorInterface;
 
 /**
  * Chain of validators for stock source link. Extension point for new validators via di configuration
+ *
+ * @api
  */
-class ValidatorChain implements StockSourceLinkValidatorInterface
+class StockSourceLinkValidatorChain implements StockSourceLinkValidatorInterface
 {
     /**
      * @var ValidationResultFactory
