+++ app/code/Magento/InventoryApi/etc/di.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <preference for="Magento\InventoryApi\Model\StockSourceLinkValidatorInterface" type="Magento\InventoryApi\Model\StockSourceLinkValidatorChain" />
+    <preference for="Magento\InventoryApi\Model\StockValidatorInterface" type="Magento\InventoryApi\Model\StockValidatorChain"/>
+    <preference for="Magento\InventoryApi\Model\SourceItemValidatorInterface" type="Magento\InventoryApi\Model\SourceItemValidatorChain"/>
+    <preference for="Magento\InventoryApi\Model\SourceValidatorInterface" type="Magento\InventoryApi\Model\SourceValidatorChain"/>
+</config>
