+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -44,7 +44,7 @@
         <plugin name="adapt_add_stock_status_to_products" type="Magento\InventoryCatalog\Plugin\CatalogInventory\Helper\Stock\AdaptAddStockStatusToProductsPlugin"/>
         <plugin name="adapt_assign_status_to_product" type="Magento\InventoryCatalog\Plugin\CatalogInventory\Helper\Stock\AdaptAssignStatusToProductPlugin"/>
     </type>
-    <type name="Magento\Inventory\Model\StockSourceLink\Validator\ValidatorChain">
+    <type name="Magento\InventoryApi\Model\StockSourceLinkValidatorChain">
         <arguments>
             <argument name="validators" xsi:type="array">
                 <item name="default_stock_default_source" xsi:type="object">
